Pull Dev into main for workflow

+++ RoguelikeFramework/Assets/Framework/Scripts/Status Effects/DamageResistance.cs
@@ -57,7 +57,7 @@ public class DamageResistance : Effect
     /*public override void OnAttacked(ref int pierce, ref int accuracy) {}*/
 
     //Called when a monster takes damage from any source, good for making effects fire upon certain types of damage
-    public override void OnTakeDamage(ref int damage, ref DamageType damageType, ref DamageSource source)
+    public override void OnTakeDamage(ref float damage, ref DamageType damageType, ref DamageSource source)
     {
         damage = damage - 3;
     }
